Refactor: 설정 탭 - 닉네임 수정 화면 Base 코드 적용
