+++ meaning-out/Controllers/Setting/EditNicknameViewController.swift
@@ -7,22 +7,12 @@
 
 import UIKit
 
-import SnapKit
-import TextFieldEffects
-
 /**
  메인 - 설정 탭 - 프로필 수정 화면
  */
-final class EditNicknameViewController: UIViewController {
+final class EditNicknameViewController: BaseViewController {
     
-    let profileImageView = UIView()
-    let profileImage = UIImageView()
-    
-    let cameraImageView = UIView()
-    let cameraImage = UIImageView()
-    
-    let nicknameField = HoshiTextField()
-    let invalidMessage = UILabel()
+    private var editNicknameView = EditNicknameView()
     
     var isValidate = false
     
@@ -35,19 +25,18 @@ final class EditNicknameViewController: UIViewController {
     }
     var profileNum = UserDefaultsManager.profile {
         didSet {
-            profileImage.image = Resource.Images.profiles[profileNum]
+            editNicknameView.profileImage.image = Resource.Images.profiles[profileNum]
         }
     }
     
     let viewModel = EditNicknameViewModel()
     
+    override func loadView() {
+        self.view = editNicknameView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        configureView()
-        configureHierarchy()
-        configureLayout()
-        configureUI()
         configureData()
         configureHandler()
         bindData()
@@ -65,7 +54,7 @@ final class EditNicknameViewController: UIViewController {
         }
         
         viewModel.outputInvalidMessage.bind { message in
-            self.invalidMessage.text = message
+            self.editNicknameView.invalidMessage.text = message
         }
         
         viewModel.outputAlertText.bind { (isValid, title, message) in
@@ -83,7 +72,7 @@ final class EditNicknameViewController: UIViewController {
         }
     }
     
-    private func configureView() {
+    override func configureViewController() {
         if !isUser {
             showAlert(title: "유효한 유저가 아니에요!", 
                       message: "온보딩 화면으로 돌아갑니다.",
@@ -93,98 +82,25 @@ final class EditNicknameViewController: UIViewController {
             return
         }
         
-        view.backgroundColor = Resource.Colors.white
         navigationItem.title = Constants.Title.edit.rawValue
-        
         addImgBarBtn(image: Resource.SystemImages.left, style: .plain, target: self, action: #selector(popViewController), type: .left)
         addTextBarBtn(title: Constants.Button.save.rawValue, style: .plain, target: self, action: #selector(saveButtonClicked), type: .right)
     }
     
-    private func configureHierarchy() {
-        cameraImageView.addSubview(cameraImage)
-        
-        profileImageView.addSubview(profileImage)
-        profileImageView.addSubview(cameraImageView)
-        
-        let subviews: [UIView] = [profileImageView, nicknameField, invalidMessage]
-        subviews.forEach { subview in
-            view.addSubview(subview)
-        }
-    }
-    
-    private func configureLayout() {
-        profileImageView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(16)
-            $0.size.equalTo(100)
-            $0.centerX.equalTo(view.safeAreaLayoutGuide)
-        }
-        
-        profileImage.snp.makeConstraints {
-            $0.top.bottom.equalTo(profileImageView)
-            $0.size.equalTo(profileImageView)
-            $0.centerX.equalTo(profileImageView)
-        }
-        
-        cameraImageView.snp.makeConstraints {
-            $0.trailing.equalTo(profileImage)
-            $0.bottom.equalTo(profileImage).inset(8)
-            $0.size.equalTo(24)
-        }
-        
-        cameraImage.snp.makeConstraints {
-            $0.center.equalTo(cameraImageView)
-            $0.size.equalTo(15)
-        }
-        
-        nicknameField.snp.makeConstraints {
-            $0.top.equalTo(profileImageView.snp.bottom).offset(24)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(24)
-            $0.height.equalTo(50)
-        }
-        
-        invalidMessage.snp.makeConstraints {
-            $0.top.equalTo(nicknameField.snp.bottom).offset(8)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(24)
-            $0.height.equalTo(30)
-        }
-    }
-    
-    private func configureUI() {
-        profileImage.backgroundColor = Resource.Colors.white
-        profileImage.clipsToBounds = true
-        profileImage.layer.cornerRadius = 50  // 추후 상수화 하기
-        profileImage.layer.borderColor = Resource.Colors.primary.cgColor
-        profileImage.layer.borderWidth = CGFloat(Constants.Integer.borderWidth.rawValue)
-        profileImage.contentMode = .scaleAspectFit
-        
-        cameraImageView.backgroundColor = Resource.Colors.primary
-        cameraImageView.layer.cornerRadius = 12  // 추후 상수화 하기
-        
-        cameraImage.image = Resource.SystemImages.camara
-        cameraImage.contentMode = .scaleAspectFit
-        cameraImage.tintColor = Resource.Colors.white
-        
-        nicknameField.setTextFieldUI(Constants.Placeholder.nickname.rawValue)
-        nicknameField.returnKeyType = .done
-        
-        invalidMessage.textColor = Resource.Colors.primary
-        invalidMessage.font = Resource.Fonts.regular13
-    }
-    
     private func configureData() {
-        profileImage.image = Resource.Images.profiles[profileNum]
-        nicknameField.text = nickname
+        editNicknameView.profileImage.image = Resource.Images.profiles[profileNum]
+        editNicknameView.nicknameField.text = nickname
     }
     
     private func configureHandler() {
         // 프로필 이미지 탭
         let profileTapped = UITapGestureRecognizer(target: self, action: #selector(profileTapped))
-        profileImageView.addGestureRecognizer(profileTapped)
-        profileImageView.isUserInteractionEnabled = true
+        editNicknameView.profileImageView.addGestureRecognizer(profileTapped)
+        editNicknameView.profileImageView.isUserInteractionEnabled = true
         
         // 닉네임 유효성 검사, 키보드 내리기
-        nicknameField.addTarget(self, action: #selector(nicknameFieldChanged), for: .editingChanged)
-        nicknameField.addTarget(self, action: #selector(keyboardDismiss), for: .editingDidEndOnExit)
+        editNicknameView.nicknameField.addTarget(self, action: #selector(nicknameFieldChanged), for: .editingChanged)
+        editNicknameView.nicknameField.addTarget(self, action: #selector(keyboardDismiss), for: .editingDidEndOnExit)
     }
     
     @objc func profileTapped() {
@@ -193,7 +109,7 @@ final class EditNicknameViewController: UIViewController {
     }
     
     @objc func nicknameFieldChanged() {
-        viewModel.inputNickname.value = nicknameField.text
+        viewModel.inputNickname.value = editNicknameView.nicknameField.text
     }
     
     @objc func keyboardDismiss() {
@@ -201,7 +117,7 @@ final class EditNicknameViewController: UIViewController {
     }
     
     @objc func saveButtonClicked() {
-        viewModel.inputNickname.value = nicknameField.text
+        viewModel.inputNickname.value = editNicknameView.nicknameField.text
         viewModel.inputOriginNickname.value = nickname
         viewModel.inputSaveButtonClicked.value = ()
     }
